modify code as recommended
