+++ paddle/fluid/operators/elementwise/elementwise_add_op.h
@@ -22,10 +22,10 @@ limitations under the License. */
 
 #include "paddle/fluid/framework/pten_utils.h"
 
-// only can include the headers in paddle/top/api dirs
+// only can include the headers in paddle/pten/include dirs
 #include "paddle/pten/api/lib/utils/tensor_utils.h"
 #include "paddle/pten/include/core.h"
-#include "paddle/pten/include/nn.h"
+#include "paddle/pten/include/math.h"
 
 namespace paddle {
 namespace operators {
