+++ paddle/fluid/operators/elementwise/elementwise_op_broadcast.cu.h
@@ -169,7 +169,11 @@ void LaunchBroadcastElementwiseCudaKernel(
     std::vector<framework::Tensor *> *outs, int axis, Functor func) {
   std::vector<const pten::DenseTensor *> pt_inputs;
   std::vector<pten::DenseTensor *> pt_outputs;
-  // *_tmp for cache DenseTensor
+  // TODO(YuanRisheng) *_tmp for cache DenseTensor, because the temporary
+  // DenseTensor obj
+  // generated by MakePtenDenseTensor can be destroyed when exits loop. *_tmp
+  // can be deleted
+  // when DenseTensor support copy constructor.
   std::vector<std::unique_ptr<pten::DenseTensor>> pt_inputs_tmp;
   std::vector<std::unique_ptr<pten::DenseTensor>> pt_outputs_tmp;
   for (auto in : ins) {
@@ -197,7 +201,11 @@ void LaunchElementwiseCudaKernel(
     std::vector<framework::Tensor *> *outs, int axis, Functor func) {
   std::vector<const pten::DenseTensor *> pt_inputs;
   std::vector<pten::DenseTensor *> pt_outputs;
-  // *_tmp for cache DenseTensor
+  // TODO(YuanRisheng) *_tmp for cache DenseTensor, because the temporary
+  // DenseTensor obj
+  // generated by MakePtenDenseTensor can be destroyed when exits loop. *_tmp
+  // can be deleted
+  // when DenseTensor support copy constructor.
   std::vector<std::unique_ptr<pten::DenseTensor>> pt_inputs_tmp;
   std::vector<std::unique_ptr<pten::DenseTensor>> pt_outputs_tmp;
   for (auto in : ins) {
