+++ paddle/pten/api/CMakeLists.txt
@@ -1,3 +1,3 @@
 add_subdirectory(lib)
 
-cc_library(pten_api SRCS all.cc DEPS linalg_api math_api creation_api manipulation_api nn_api)
+cc_library(pten_api SRCS all.cc DEPS linalg_api math_api creation_api manipulation_api)
