+++ paddle/pten/api/lib/CMakeLists.txt
@@ -4,4 +4,3 @@ cc_library(math_api SRCS math.cc DEPS pten)
 cc_library(linalg_api SRCS linalg.cc DEPS pten)
 cc_library(creation_api SRCS creation.cc DEPS pten)
 cc_library(manipulation_api SRCS manipulation.cc DEPS pten)
-cc_library(nn_api SRCS nn.cc DEPS pten)
