+++ paddle/pten/kernels/cpu/math.cc
@@ -14,13 +14,16 @@
 
 #include "paddle/pten/kernels/cpu/math.h"
 
+#include "paddle/pten/kernels/functions/cpu/elementwise.h"
 #include "paddle/pten/kernels/functions/eigen/mean.h"
 #include "paddle/pten/kernels/functions/eigen/scale.h"
 #include "paddle/pten/kernels/functions/eigen/sign.h"
+#include "paddle/pten/kernels/functions/general/elementwise_functor.h"
 
 // See Note [ Why still include the fluid headers? ]
 #include "paddle/fluid/framework/eigen.h"
 #include "paddle/fluid/platform/bfloat16.h"
+#include "paddle/fluid/platform/complex.h"
 
 namespace pten {
 
@@ -61,10 +64,34 @@ void ScaleHost(const CPUContext& dev_ctx,
                               out);
 }
 
+template <typename T>
+void ElementwiseAdd(const CPUContext& dev_ctx,
+                    const DenseTensor& x,
+                    const DenseTensor& y,
+                    int axis,
+                    DenseTensor* out) {
+  if (x.dims() == y.dims()) {
+    SameDimsElementwiseCompute<general::SameDimsAddFunctor<CPUContext, T>>()(
+        dev_ctx, x, y, out);
+  } else {
+    auto x_dims = x.dims();
+    auto y_dims = y.dims();
+    if (x_dims.size() >= y_dims.size()) {
+      ElementwiseCompute<general::AddFunctor<T>, T>(
+          dev_ctx, x, y, axis, general::AddFunctor<T>(), out);
+    } else {
+      ElementwiseCompute<general::InverseAddFunctor<T>, T>(
+          dev_ctx, x, y, axis, general::InverseAddFunctor<T>(), out);
+    }
+  }
+}
 }  // namespace pten
 
 // TODO(chenweihang): replace by better impl
 PT_REGISTER_MODULE(MathCPU);
+
+using complex64 = ::paddle::platform::complex<float>;
+using complex128 = ::paddle::platform::complex<double>;
 
 // NOTE(chenweihang): using bfloat16 will cause redefine with xpu bfloat16
 // using bfloat16 = ::paddle::platform::bfloat16;
@@ -97,3 +124,13 @@ PT_REGISTER_KERNEL("scale.host",
                    int64_t) {
   kernel->InputAt(1).SetBackend(pten::Backend::CPU);
 }
+PT_REGISTER_KERNEL("elementwise_add",
+                   CPU,
+                   ANY,
+                   pten::ElementwiseAdd,
+                   float,
+                   double,
+                   int,
+                   int64_t,
+                   complex64,
+                   complex128) {}
