+++ paddle/pten/tests/kernels/test_elementwise_dev_api.cc
@@ -15,16 +15,16 @@ limitations under the License. */
 #include <gtest/gtest.h>
 #include <memory>
 
-#include "paddle/pten/include/nn.h"
+#include "paddle/pten/include/math.h"
 
 #include "paddle/pten/api/lib/utils/allocator.h"
 #include "paddle/pten/core/dense_tensor.h"
 #include "paddle/pten/core/kernel_registry.h"
 
-PT_DECLARE_MODULE(NnCPU);
+PT_DECLARE_MODULE(MathCPU);
 
 #if defined(PADDLE_WITH_CUDA) || defined(PADDLE_WITH_HIP)
-PT_DECLARE_MODULE(LinalgCUDA);
+PT_DECLARE_MODULE(MathCUDA);
 #endif
 
 namespace framework = paddle::framework;
